+++ tests/defaults_file_test.py
@@ -14,18 +14,32 @@ class DefaultsFileTest(unittest.TestCase):
     @tempdir()
     def test_set_defaults_creates_file_if_nonexistent(self):
         self.assertIs(os.path.isfile(DEFAULTS_FILE), False)
-        ParserDefinition('test_script').set_defaults()
+        ParserDefinition.create('test_script.py').set_defaults()
         self.assertIs(os.path.isfile(DEFAULTS_FILE), True)
 
     @tempdir()
+    def test_set_defaults_nested(self):
+        parser_def = argutil.ParserDefinition.create('test_script.py')
+        parser_def.set_defaults(**{"command1.command2": "value"})
+        expected = {
+            "test_script": {
+                "command1": {
+                    "command2": "value"
+                }
+            }
+        }
+        actual = parser_def.load(parser_def.defaults_file)
+        self.assertDictEqual(actual, expected)
+
+    @tempdir()
     def test_get_defaults_nonexistent_defaults_file(self):
-        parser_def = ParserDefinition('test_script')
+        parser_def = ParserDefinition.create('test_script.py')
         self.assertDictEqual(parser_def.get_defaults(), {})
 
     @tempdir()
     def test_get_defaults(self):
         expected = {}
-        parser_def = ParserDefinition('test_script.py')
+        parser_def = ParserDefinition.create('test_script.py')
         parser_def.set_defaults()
         self.assertDictEqual(
             parser_def.get_defaults(),
@@ -38,7 +52,7 @@ class DefaultsFileTest(unittest.TestCase):
             'foo': 'bar',
             'bar2': 'foo2'
         }
-        parser_def = ParserDefinition('test_script.py')
+        parser_def = ParserDefinition.create('test_script.py')
         parser_def.set_defaults(foo='bar', bar2='foo2')
         self.assertDictEqual(
             parser_def.get_defaults(),
@@ -48,21 +62,21 @@ class DefaultsFileTest(unittest.TestCase):
     @tempdir()
     def test_get_defaults_nonexistent_module(self):
         expected = {}
-        parser_def = ParserDefinition('test_script.py')
+        parser_def = ParserDefinition.create('test_script.py')
         parser_def.set_defaults(foo='bar')
-        parser_def2 = ParserDefinition('other_script.py')
+        parser_def2 = ParserDefinition.create('other_script.py')
         self.assertDictEqual(parser_def2.get_defaults(), expected)
 
     @tempdir()
     def test_config_fetch_no_defaults(self):
-        parser_def = ParserDefinition('test_script.py')
+        parser_def = ParserDefinition.create('test_script.py')
         current = parser_def.config()
         expected = []
         self.assertEqual(current, expected)
 
     @tempdir()
     def test_config_fetch_has_defaults(self):
-        parser_def = ParserDefinition('test_script.py')
+        parser_def = ParserDefinition.create('test_script.py')
         parser_def.set_defaults(
             foo='bar',
             bar='foo'
@@ -75,8 +89,20 @@ class DefaultsFileTest(unittest.TestCase):
         self.assertCollectionEqual(current, expected)
 
     @tempdir()
+    def test_config_has_nested_defaults(self):
+        parser_def = ParserDefinition.create('test_script.py')
+        parser_def.set_defaults(**{
+            "foo.bar.fee": "bar"
+        })
+        current = parser_def.config()
+        expected = [
+            'foo.bar.fee=bar'
+        ]
+        self.assertCollectionEqual(current, expected)
+
+    @tempdir()
     def test_config_single_property(self):
-        parser_def = ParserDefinition('test_script.py')
+        parser_def = ParserDefinition.create('test_script.py')
         parser_def.config(['foo=bar'])
         expected = {'foo': 'bar'}
         self.assertDictEqual(
@@ -86,7 +112,7 @@ class DefaultsFileTest(unittest.TestCase):
 
     @tempdir()
     def test_config_multiple_properties(self):
-        parser_def = ParserDefinition('test_script.py')
+        parser_def = ParserDefinition.create('test_script.py')
         parser_def.config(
             [
                 "a=1",
@@ -118,7 +144,7 @@ class DefaultsFileTest(unittest.TestCase):
 
     @tempdir()
     def test_config_list_property(self):
-        parser_def = ParserDefinition('test_script.py')
+        parser_def = ParserDefinition.create('test_script.py')
         parser_def.config(
             ['foo=[1, "2", \'3\', true, false, none, null]']
         )